feat: add keycodes and can increase and downcrease z_divisor

+++ isometric.c
@@ -6,7 +6,7 @@
 /*   By: gfrancoi <gfrancoi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/31 20:05:48 by gfrancoi          #+#    #+#             */
-/*   Updated: 2025/04/23 10:13:36 by gfrancoi         ###   ########.fr       */
+/*   Updated: 2025/04/23 11:56:40 by gfrancoi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ static void	apply_iso_to_point(t_fdf *fdf, t_point3 *src, t_point2 *dst)
 	sin_a = sin(M_PI_4);
 	x = src->v.axis[X];
 	y = src->v.axis[Y];
-	z = src->v.axis[Z] * fdf->projection.z_divisor;
+	z = (float)(src->v.axis[Z]) * fdf->projection.z_divisor;
 	dst->v.axis[X] = (x - y) * cos_a * fdf->projection.default_scale
 		+ fdf->projection.offsets[X];
 	dst->v.axis[Y] = (x + y) * sin_a * fdf->projection.default_scale
